I had to remove "BASE_MODULE_VERSION" since it was not next to "EPICS_BASE_VER" Below: is from EPICS BASE 3.15.4 release notes ============================================================================================= Additional RELEASE file checks
An additional check has been added at build-time for the contents of the configure/RELEASE file(s), which will mostly only affect users of the Debian EPICS packages published by NSLS-2. Support modules may share an install path, but all such modules must be listed adjacent to each other in any RELEASE files that point to them. For example the following will fail the new checks:

AUTOSAVE = /usr/lib/epics
ASYN = /home/mdavidsaver/asyn
EPICS_BASE = /usr/lib/epics
giving the compile-time error

This application's RELEASE file(s) define
	EPICS_BASE = /usr/lib/epics
after but not adjacent to
	AUTOSAVE = /usr/lib/epics
Module definitions that share paths must be grouped together.
Either remove a definition, or move it to a line immediately
above or below the other(s).
Any non-module definitions belong in configure/CONFIG_SITE.
In many cases such as the one above the order of the AUTOSAVE and ASYN lines can be swapped to let the checks pass, but if the AUTOSAVE module depended on ASYN and hence had to appear before it in the list this error indicates that AUTOSAVE should also be built in its own private area; a shared copy would likely be incompatible with the version of ASYN built in the home directory.